#154 add postSpecificTests
